fix state machine of hash join
