+++ query/src/pipelines/new/processors/mod.rs
@@ -47,7 +47,7 @@ pub use sources::SyncSourcer;
 pub use transforms::AggregatorParams;
 pub use transforms::AggregatorTransformParams;
 pub use transforms::BlockCompactor;
-pub use transforms::ChainHashTable;
+pub use transforms::ChainingHashTable;
 pub use transforms::ExpressionTransform;
 pub use transforms::HashJoinState;
 pub use transforms::ProjectionTransform;
